MN-567: fix test, move it out from a slowtest tag
